Refactoring exported interfaces, reverting props
