--- src/Components/Types/AppBarType.tsx
@@ -1,9 +1,0 @@
-type AppBarPropType = {
-  geocoderContainerRef: any;
-  toggleGuide: () => void;
-  map: any;
-  filterApplied: boolean;
-  clearFilters: Function;
-};
-
-export default AppBarPropType;
