+++ src/Components/Types/SearchFilterType.tsx
@@ -1,0 +1,6 @@
+type SearchFilterType = {
+  is_ordering_tests_only_for_those_who_meeting_criteria: boolean;
+  is_collecting_samples: boolean;
+};
+
+export default SearchFilterType;
