--- src/Components/Types/SidebarPropType.tsx
@@ -1,6 +1,0 @@
-type SidebarPropType = {
-  drawerOpen: boolean;
-  toggleFilter: Function;
-};
-
-export default SidebarPropType;
