+++ src/utils/fetchPins.tsx
@@ -1,8 +1,8 @@
-import { SearchFilters } from '../App';
+import SearchFilterType from '../Components/Types/SearchFilterType';
 
 const { REACT_APP_API_ENDPOINT } = process.env;
 
-const fetchPins = async (searchFilters: SearchFilters) => {
+const fetchPins = async (searchFilters: SearchFilterType) => {
   const endpoint = `${REACT_APP_API_ENDPOINT}/api/v1/location`;
   const results: Response = await fetch(endpoint);
   const data = await results.json();
@@ -12,7 +12,7 @@ const fetchPins = async (searchFilters: SearchFilters) => {
     let ret: boolean = true;
 
     for (let i = 0; i < keys.length; i++) {
-      const filter: keyof SearchFilters = keys[i] as keyof SearchFilters;
+      const filter: keyof SearchFilterType = keys[i] as keyof SearchFilterType;
       if (searchFilters[filter] && place[filter] !== true) {
         ret = false;
       }
